Move schedulers from examples.
The production ready schedulers are moved from the examples
directory and built into a new library.
