+++ sprokit/src/schedulers/registration.cxx
@@ -1,5 +1,5 @@
 /*ckwg +29
- * Copyright 2011-2012 by Kitware, Inc.
+ * Copyright 2016 by Kitware, Inc.
  * All rights reserved.
  *
  * Redistribution and use in source and binary forms, with or without
@@ -28,25 +28,37 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-#ifndef SPROKIT_SCHEDULERS_EXAMPLES_SCHEDULERS_REGISTRATION_H
-#define SPROKIT_SCHEDULERS_EXAMPLES_SCHEDULERS_REGISTRATION_H
+#include "sync_scheduler.h"
+#include "thread_per_process_scheduler.h"
 
-#include "examples-config.h"
+#include <sprokit/pipeline/scheduler_registry.h>
+#include <schedulers/schedulers_export.h>
 
 /**
- * \file examples/registration.h
+ * \file registration.cxx
  *
  * \brief Register schedulers for use.
  */
-
 extern "C"
+SCHEDULERS_EXPORT void register_schedulers();
+
+using namespace sprokit;
+
+void
+register_schedulers()
 {
+  static scheduler_registry::module_t const module_name = scheduler_registry::module_t("schedulers");
 
-/**
- * \brief Register schedulers.
- */
-SPROKIT_SCHEDULERS_EXAMPLES_EXPORT void register_schedulers();
+  scheduler_registry_t const registry = scheduler_registry::self();
 
+  if (registry->is_module_loaded(module_name))
+  {
+    return;
+  }
+
+  registry->register_scheduler("sync", "Run the pipeline synchronously", create_scheduler<sync_scheduler>);
+  registry->register_scheduler("thread_per_process", "Run each process in its own thread",
+                               create_scheduler<thread_per_process_scheduler>);
+
+  registry->mark_module_as_loaded(module_name);
 }
-
-#endif // SPROKIT_SCHEDULERS_EXAMPLES_SCHEDULERS_REGISTRATION_H
